pyproject.toml
[project]
name = "wikipedia-pageview-analytics"
version = "0.1.0"
description = "wikipedia page view."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "pyspark>=4.1.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
